Remove V2 from blocks and replace old blocks with new ones.
