+++ .github/coverage-config.json
@@ -42,8 +42,6 @@
     "src/helpers/urlrewrite.js",
     "src/helpers/interactivity.js",
     "src/blocks/events-list/events-list.js",
-    "src/blocks/online-event/online-event.js",
-    "src/blocks/venue/venue.js",
     "src/blocks/rsvp-response/rsvp-manager.js"
   ],
   "allowLowCoverage": [
