+++ src/blocks/online-event-link/block.json
@@ -7,7 +7,7 @@
 	"category": "gatherpress",
 	"icon": "admin-links",
 	"description": "Display an editable link for online events. Link URL comes from event meta, text is customizable.",
-	"ancestor": ["gatherpress/online-event-v2"],
+	"ancestor": ["gatherpress/online-event"],
 	"usesContext": ["postId", "queryId"],
 	"attributes": {
 		"linkText": {
