--- src/blocks/online-event-v2/block.json
@@ -1,28 +1,0 @@
-{
-	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 3,
-	"name": "gatherpress/online-event-v2",
-	"version": "2.0.0",
-	"title": "Online Event (v2)",
-	"category": "gatherpress",
-	"icon": "video-alt2",
-	"description": "Container block for online event display with icon and link.",
-	"usesContext": ["postId", "queryId"],
-	"providesContext": {
-		"postId": "postId"
-	},
-	"attributes": {
-		"postId": {
-			"type": "number"
-		}
-	},
-	"supports": {
-		"gatherpress": {
-			"blockGuard": true,
-			"postIdOverride": true
-		},
-		"html": false
-	},
-	"textdomain": "gatherpress",
-	"editorScript": "file:./index.js"
-}
