--- src/blocks/online-event-v2/index.js
@@ -1,31 +1,0 @@
-/**
- * External dependencies.
- */
-import { registerBlockType } from '@wordpress/blocks';
-import { InnerBlocks, useBlockProps } from '@wordpress/block-editor';
-
-/**
- * Internal dependencies.
- */
-import edit from './edit';
-import metadata from './block.json';
-
-/**
- * Register the GatherPress Online Event v2 block.
- *
- * Container block for online event display with icon and link.
- *
- * @since 1.0.0
- *
- * @return {void}
- */
-registerBlockType( metadata, {
-	edit,
-	save: () => {
-		return (
-			<div { ...useBlockProps.save() }>
-				<InnerBlocks.Content />
-			</div>
-		);
-	},
-} );
