+++ src/blocks/online-event/block.json
@@ -1,24 +1,28 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
-	"apiVersion": 2,
+	"apiVersion": 3,
 	"name": "gatherpress/online-event",
-	"version": "1.0.1",
+	"version": "2.0.0",
 	"title": "Online Event",
 	"category": "gatherpress",
 	"icon": "video-alt2",
-	"example": {},
-	"description": "Utilized for virtual events, offering the capability to share a direct link to an event.",
+	"description": "Container block for online event display with icon and link.",
+	"usesContext": ["postId", "queryId"],
+	"providesContext": {
+		"postId": "postId"
+	},
 	"attributes": {
-		"blockId": {
-			"type": "string"
+		"postId": {
+			"type": "number"
 		}
 	},
 	"supports": {
+		"gatherpress": {
+			"blockGuard": true,
+			"postIdOverride": true
+		},
 		"html": false
 	},
 	"textdomain": "gatherpress",
-	"editorScript": "file:./index.js",
-	"style": "file:./style-index.css",
-	"viewScript": "file:./online-event.js",
-	"render": "file:./render.php"
+	"editorScript": "file:./index.js"
 }
