+++ src/blocks/online-event/index.js
@@ -2,6 +2,8 @@
  * External dependencies.
  */
 import { registerBlockType } from '@wordpress/blocks';
+import { InnerBlocks, useBlockProps } from '@wordpress/block-editor';
+
 /**
  * Internal dependencies.
  */
@@ -9,12 +11,9 @@ import edit from './edit';
 import metadata from './block.json';
 
 /**
- * Register the GatherPress Online Event block.
+ * Register the GatherPress Online Event v2 block.
  *
- * This code registers the GatherPress Online Event block in the WordPress block editor.
- * It uses the block metadata from the 'block.json' file and associates it with the
- * edit component for rendering in the editor. The 'save' function is set to null as
- * the block doesn't have a front-end representation and is only used in the editor.
+ * Container block for online event display with icon and link.
  *
  * @since 1.0.0
  *
@@ -22,5 +21,11 @@ import metadata from './block.json';
  */
 registerBlockType( metadata, {
 	edit,
-	save: () => null,
+	save: () => {
+		return (
+			<div { ...useBlockProps.save() }>
+				<InnerBlocks.Content />
+			</div>
+		);
+	},
 } );
