--- src/blocks/venue-v2/render.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * Render Venue v2 block.
- *
- * @package GatherPress\Core
- * @since 1.0.0
- */
-
-// Exit if accessed directly.
-defined( 'ABSPATH' ) || exit; // @codeCoverageIgnore
-
-// Render the block with inner blocks content.
-printf(
-	'<div %s>%s</div>',
-	get_block_wrapper_attributes(), // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-	$content // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-);
