--- src/blocks/venue/style.scss
@@ -1,29 +1,0 @@
-.gatherpress-venue {
-	display: flex;
-	flex-direction: column;
-	gap: 0.5rem;
-
-	&__gap {
-		gap: 2rem;
-	}
-
-	&__icon {
-		align-items: flex-start;
-		display: flex;
-		padding-top: 0.4rem;
-	}
-
-	&__item {
-		display: flex;
-		gap: 1rem;
-	}
-
-	&__name {
-		font-size: var(--wp--preset--font-size--medium);
-	}
-
-	&__row {
-		display: flex;
-		justify-content: flex-start;
-	}
-}
